requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
gunicorn==21.2.0
qrcode==7.4.2
Pillow==10.4.0
